Sample of UK manifesto sentences 2010 crowd-annotated for immigration

A corpus of sentences sampled from from publicly available party manifestos from the United Kingdom from the 2010 election. Each sentence has been rated in terms of its classification as pertaining to immigration or not and then on a scale of favorability or not toward open immigration policy (as the mean score of crowd coders on a scale of -1 (favours open immigration policy), 0 (neutral), or 1 (anti-immigration).

The sentences were sampled from the corpus used in [Benoit et al. (2016)](https://doi.org/10.1017/S0003055416000058), which contains more information on the crowd-sourced annotation approach.

Usage

data_corpus_manifsentsUK2010sample

Format

The corpus consists of 155 sentences randomly sampled from the party manifestos, with an attempt to balance the sentencs according to their categorisation as pertaining to immigration or not, as well as by party. The corpus contains the following document-level variables:

party

factor; abbreviation of the party that wrote the manifesto.

partyname

factor; party that wrote the manifesto.

year

integer; 4-digit year of the election.

crowd_immigration_label

Factor indicating whether the majority of crowd workers labelled a sentence as referring to immigration or not. The variable has missing values (`NA`) for all non-annotated manifestos.

crowd_immigration_mean

numeric; the direction of statements coded as "Immigration" based on the aggregated crowd codings. The variable is the mean of the scores assigned by workers who coded a sentence and who allocated the sentence to the "Immigration" category. The variable ranges from -1 (Favorable and open immigration policy) to +1 ("Negative and closed immigration policy").

crowd_immigration_n

integer; the number of coders who contributed to the mean score `crowd_immigration_mean`.

A [corpus][quanteda::corpus] object.

References

Benoit, K., Conway, D., Lauderdale, B.E., Laver, M., & Mikhaylov, S. (2016). [Crowd-sourced Text Analysis: Reproducible and Agile Production of Political Data](https://doi.org/10.1017/S0003055416000058). *American Political Science Review*, 100,(2), 278–295.

Examples

if (FALSE) { # \dontrun{
library(quanteda)

immigration_instructions <- "This task involves reading sentences from
political texts from the 2010 UK general election, and judging whether
these statements deal with immigration policy. Each sentence may or may
not be related to immigration policy.

First, you will read a short section from a party manifesto. For the
sentence highlighted in red, enter your best judgment about whether it
refers to some aspect of immigration policy, or not. Most sentences will
not relate to immigration policy - it is your job to find and rate those
that do. If the sentence does not refer to immigration policy, you should
select 'Not immigration policy' and proceed directly to the next sentence.
If the sentence does refer to immigration policy, you should indicate this
by checking this option.

If you indicate that the sentence is immigration policy, you will be asked
to give your best judgment of the policy position on immigration being
expressed in the sentence. This will range from a very open and favourable
position on immigration, to a very closed and negative stance on
immigration. These are coded on a five-point scale, with a neutral position
(neither favouring nor opposing) immigration lying in the middle."

immigration_description <- "What is 'immigration policy'?

Immigration policy relates to all government policies, laws, regulations,
and practices that deal with the free travel of foreign persons across the
country's borders, especially those that intend to live, work, or seek
legal protection (asylum) in that country. Examples of specific policies
that pertain to immigration include the regulation of: work permits for
foreign nationals; residency permits for foreign nationals; asylum seekers
and their treatment; requirements for acquiring citizenship; illegal
immigrants and migrant workers (and their families) living or working
illegally in the country.

It also includes favorable or unfavorable general statements about
immigrants or immigration policy, such as statements indicating that
immigration has been good for a country, or that immigrants have forced
local people out of jobs, etc."

immigration_scale <- "Pro-immigration policies (a value of -1)

Examples of 'pro' immigration positions include: Positive statements about
the benefits of immigration, such as economic or cultural benefits;
Statements about the moral obligation to welcome asylum seekers; Policies
that would improve conditions for asylum seekers and their families; Urging
an increase the number of work permits for foreign nationals; Making it
possible for illegal immigrants to obtain a legal status or even
citizenship; Reducing barriers to immigration generally.

Anti-immigration policies (a value of 1)

Examples of 'anti' immigration positions include: Negative statements about
consequences of immigration, such as job losses, increased crime, or
destruction of national culture; Arguments about asylum seekers abusing the
system; Policies to deport asylum seekers and their families; Urging
restrictions on the number of work permits for foreign nationals, including
points systems; Deporting illegal immigrants and their families; Increasing
barriers to immigration generally.

Neutral immigration policy statements (a value of 0)

Examples of neutral statements about immigration policy: Advocating a
balanced approach to the problem; Statements about administrative capacity
for handling immigration or asylum seekers; Statements that do not take a
pro- or anti-immigration stance generally, despite making some statement
about immigration."

# define a sentiment task
task_immigration <- task(
  name = "Immigration policy",
  system_prompt = immigration_instructions,
  type_def = type_object(
    llm_immigration_label = type_enum(c("Immigration", "Not immmigration")),
    llm_immigration_scale = type_integer(immigration_scale),
  )
)

result <- data_corpus_manifsentsUK2010sample %>%
  annotate(task_immigration,
           chat_fn = chat_google_gemini, model = "gemini-2.5-flash")

result_combined <- cbind(
  result,
  data.frame(crowd_immigration_label = data_corpus_manifsentsUK2010sample$crowd_immigration_label)
)

# compute agreement
agreement(result_combined, unit_id_col = "id",
          coder_cols = c("llm_immigration_label", "crowd_immigration_label"))

# confusion matrix
table(tmp$llm_immigration_label, tmp$crowd_immigration_label)
} # }